Drupal to Blogger script
To use this script first create your blog into Blogger, create a test
posts
and export it to XML. Then run this php script and copy paste the
output towards
the bottom of the XML, where your test post is located.
Save the file and import it again in Blogger. It usually takes some
time,
but in the end you get the message that everything is imported
correctly.